gcp/requirements.txt
tensorflow
pillow
numpy
flask
google-cloud-storage